Revert "update RubyGems to 2.5.1 (same as in MRI 2.3.0)"
This reverts commit 604ac226a71e9921d19f11a89e7a7f482a33267b.

+++ lib/ruby/stdlib/rubygems/ext/ext_conf_builder.rb
@@ -35,12 +35,7 @@ class Gem::Ext::ExtConfBuilder < Gem::Ext::Builder
         begin
           run cmd, results
         ensure
-          if File.exist? 'mkmf.log'
-            results << "To see why this extension failed to compile, please check" \
-              " the mkmf.log which can be found here:\n"
-            results << "  " + File.join(dest_path, 'mkmf.log') + "\n"
-            FileUtils.mv 'mkmf.log', dest_path
-          end
+          FileUtils.mv 'mkmf.log', dest_path if File.exist? 'mkmf.log'
           siteconf.unlink
         end
